refactor: extract pages to be at the same level as components
